requirements.txt
pandas
numpy
requests
python-telegram-bot==21.4
python-dotenv
